requirements.txt
openai
fastapi
black
isort
websockets
pydantic
langchain
uvicorn
jinja2
faiss-cpu
bs4
unstructured
libmagic
tiktoken
python-dotenv